pyproject.toml
[project]
name = "new-admin-panel-sprint-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==4.2.11",
    "django-debug-toolbar>=6.0.0",
    "django-split-settings>=1.3.2",
    "flake8==6.1.0",
    "psycopg>=3.2.9",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
fixable = ["ALL"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-single-line = false
lines-after-imports = 2
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]
known-first-party = ["movies_admin", "sqlite_to_postgres"]
